Refactor apis command (#541)
