+++ src/commands.ts
@@ -4,20 +4,17 @@ import { readFileSync } from 'fs';
  */
 import chalk from 'chalk';
 import * as pluralize from 'pluralize';
-import { PUBLIC_ADVANCED_SERVICES, SCRIPT_TYPES } from './apis';
+import { SCRIPT_TYPES } from './apis';
 import {
   enableAppsScriptAPI,
-  enableOrDisableAPI,
   getFunctionNames,
 } from './apiutils';
 import {
   authorize,
-  discovery,
   drive,
   getLocalScript,
   loadAPICredentials,
   script,
-  serviceUsage,
 } from './auth';
 import { fetchProject, hasProject, writeProjectFiles } from './files';
 import {
@@ -30,10 +27,8 @@ import { URL } from './urls';
 import {
   ERROR,
   LOG,
-  PROJECT_MANIFEST_BASENAME,
   checkIfOnline,
   getDefaultProjectName,
-  getProjectId,
   getProjectSettings,
   hasOauthClientSettings,
   logError,
@@ -41,9 +36,7 @@ import {
   spinner,
 } from './utils';
 
-const open = require('opn');
 const inquirer = require('inquirer');
-const padEnd = require('string.prototype.padend');
 
 // setup inquirer
 const prompt = inquirer.prompt;
@@ -362,109 +355,4 @@ https://www.googleapis.com/auth/presentations
     }
   }
   await runFunction(functionName, params);
-};
-
-/**
- * Acts as a router to apis subcommands
- * Calls functions for list, enable, or disable
- * Otherwise returns an error of command not supported
- */
-export const apis = async (options: { open?: string }) => {
-  await loadAPICredentials();
-  const subcommand: string = process.argv[3]; // clasp apis list => "list"
-  const serviceName = process.argv[4]; // clasp apis enable drive => "drive"
-
-  // clasp apis --open
-  if (options.open) {
-    const apisUrl = URL.APIS(await getProjectId());
-    console.log(apisUrl);
-    return open(apisUrl, { wait: false });
-  }
-
-  // The apis subcommands.
-  const command: { [key: string]: Function } = {
-    enable: async () => {
-      enableOrDisableAPI(serviceName, true);
-    },
-    disable: async () => {
-      enableOrDisableAPI(serviceName, false);
-    },
-    list: async () => {
-      await checkIfOnline();
-      /**
-       * List currently enabled APIs.
-       */
-      console.log('\n# Currently enabled APIs:');
-      const projectId = await getProjectId(); // will prompt user to set up if required
-      const MAX_PAGE_SIZE = 200; // This is the max page size according to the docs.
-      const list = await serviceUsage.services.list({
-        parent: `projects/${projectId}`,
-        filter: 'state:ENABLED',
-        pageSize: MAX_PAGE_SIZE,
-      });
-      const serviceList = list.data.services || [];
-      if (serviceList.length >= MAX_PAGE_SIZE) {
-        console.log('Uh oh. It looks like Grant did not add pagination. Please create a bug.');
-      }
-
-      // Filter out the disabled ones. Print the enabled ones.
-      const enabledAPIs = serviceList.filter((service: any) => {
-        return service.state === 'ENABLED';
-      });
-      for (const enabledAPI of enabledAPIs) {
-        if (enabledAPI.config && enabledAPI.config.documentation) {
-          const name = enabledAPI.config.name || 'Unknown name.';
-          console.log(`${name.substr(0, name.indexOf('.'))} - ${enabledAPI.config.documentation.summary}`);
-        }
-      }
-
-      /**
-       * List available APIs.
-       */
-      console.log('\n# List of available APIs:');
-      const { data } = await discovery.apis.list({
-        preferred: true,
-      });
-      const services = data.items || [];
-      // Only get the public service IDs
-      const PUBLIC_ADVANCED_SERVICE_IDS = PUBLIC_ADVANCED_SERVICES.map(
-        advancedService => advancedService.serviceId,
-      );
-
-      // Merge discovery data with public services data.
-      const publicServices = [];
-      for (const publicServiceId of PUBLIC_ADVANCED_SERVICE_IDS) {
-        const service: any = services.find((s: any) => s.name === publicServiceId);
-        // for some reason 'youtubePartner' is not in the api list.
-        if (service && service.id && service.description) {
-          publicServices.push(service);
-        }
-      }
-
-      // Sort the services based on id
-      publicServices.sort((a: any, b: any) => {
-        if (a.id < b.id) return -1;
-        if (a.id > b.id) return 1;
-        return 0;
-      });
-
-      // Format the list
-      for (const api of publicServices) {
-        console.log(`${padEnd(api.name, 25)} - ${padEnd(api.description, 60)}`);
-      }
-    },
-    undefined: () => {
-      command.list();
-
-      console.log(`# Try these commands:
-- clasp apis list
-- clasp apis enable slides
-- clasp apis disable slides`);
-    },
-  };
-  if (command[subcommand]) {
-    command[subcommand]();
-  } else {
-    logError(null, ERROR.COMMAND_DNE('apis ' + subcommand));
-  }
 };
